Various small code improvements

+++ application/modules/events/controllers/admin/Index.php
@@ -59,11 +59,9 @@ class Index extends \Ilch\Controller\Admin
         $this->getLayout()->getAdminHmenu()
             ->add($this->getTranslator()->trans('menuEvents'), ['action' => 'index']);
 
-        if ($this->getRequest()->getPost('check_entries')) {
-            if ($this->getRequest()->getPost('action') === 'delete') {
-                foreach ($this->getRequest()->getPost('check_entries') as $eventId) {
-                    $eventMapper->delete($eventId);
-                }
+        if ($this->getRequest()->getPost('check_entries') && $this->getRequest()->getPost('action') === 'delete') {
+            foreach ($this->getRequest()->getPost('check_entries') as $eventId) {
+                $eventMapper->delete($eventId);
             }
         }
